reorganize config, add option to specify plugin dependencies
