+++ components/bpmn-q/modeler-component/editor/config/GeneralTab.js
